Refactor crate generators to use repo2rocrate library
