+++ lifemonitor/api/models/rocrate/generators.py
@@ -19,32 +19,20 @@
 # SOFTWARE.
 
 
-import glob
 import inspect
 import logging
-import os
 from importlib import import_module
 from pathlib import Path
-from typing import Dict, List
+from typing import List
+
 
 # set module level logger
 logger = logging.getLogger(__name__)
 
 
-__modules_files__ = glob.glob(os.path.join(os.path.dirname(__file__), "*.py"))
-__modules__ = {os.path.basename(f)[:-3]: '{}.{}'.format(__name__, os.path.basename(f)[:-3])
-               for f in __modules_files__ if os.path.isfile(f) and not f.endswith('__init__.py')}
-logger.debug("Loaded modules of RO-Crate generators: %r", __modules__)
-
-
-class GenCrateConfig:
-    def __init__(self, opts: Dict) -> None:
-        for k, v in opts.items():
-            setattr(self, k, v)
-
-
 def get_supported_workflow_types() -> List[str]:
-    return list(__modules__.keys())
+    from repo2rocrate.cli import GEN_MAP
+    return list(GEN_MAP.keys())
 
 
 def generate_crate(workflow_type: str, workflow_version: str,
@@ -62,31 +50,28 @@ def generate_crate(workflow_type: str, workflow_version: str,
     #       The current implementation only supports Snakemake and Github CI
     cfg = {
         "root": Path(local_repo_path),
-        "output": local_repo_path,
         "repo_url": repo_url,
         "version": workflow_version,
         "license": license,
         "ci_workflow": kwargs.get('ci_workflow', 'main.yml'),
         "lang_version": kwargs.get('lan_version', '0.6.5')
     }
-    opts = GenCrateConfig(cfg)
     logger.warning("Config: %r", cfg)
-    make_crate(opts)
+    crate = make_crate(**cfg)
+    crate.write(local_repo_path)
+    return crate
 
 
 def get_crate_generator(workflow_type: str):
-    mod_name = __modules__.get(workflow_type, None)
-    if not mod_name:
-        raise NotImplementedError('No RO-Crate generator for workflow type "%s"', workflow_type)
     try:
-        mod = import_module(mod_name)
+        mod = import_module(f"repo2rocrate.{workflow_type}")
         make_crate = getattr(mod, "make_crate")
         logger.debug("Found make_crate: %r", make_crate)
         if not inspect.isfunction(make_crate):
             logger.warning("'make_crate' in %r is not a function")
         return make_crate
     except ModuleNotFoundError:
-        logger.error("ModuleNotFoundError: Unable to load module %s", mod_name)
+        raise NotImplementedError('No RO-Crate generator for workflow type "%s"' % workflow_type)
     except AttributeError:
         logger.error("AttributeError: Unable to find function make_crate on module %s", mod)
     return None
